requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
huggingface-hub>=0.16.0
langdetect>=1.0.9
requests>=2.28.0
pydub>=0.25.1
python-multipart>=0.0.5